requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==4.4.2
Flask==3.0.2
Flask-WTF==1.2.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==1.22.2
opencv-python==4.5.5.62
pillow==10.2.0
proglog==0.1.10
pycparser==2.21
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0